+++ internal/accrual/model/order.go
@@ -12,6 +12,6 @@ const (
 type Order struct {
 	Number    uint64    `db:"number" json:"number" required:"true"`
 	Status    string    `db:"status" json:"status"`
-	Accrual   *uint64   `db:"accrual" json:"accrual"`
+	Accrual   *float64  `db:"accrual" json:"accrual"`
 	CreatedAt time.Time `db:"created_at" json:"-"`
 }
